Extract RestInfo objects into OpenAcdUtilities for reuse among APIs
